dev(ierror): finished the error.go and utils.go tests
