+++ pkg/ierrors/utils.go
@@ -1,30 +1,8 @@
 package ierrors
 
-import "errors"
-
-// TODO add better description
-// Is Compares errors
-func Is(source, target error) bool {
-
-	// check if is another type of error inside the error stack
-	if errors.Is(source, target) {
-		return true
-	}
-
-	// converts target to ierror structure, if possible
-	t, ok := target.(*ierror)
-	if !ok {
-		return false
-	}
-
-	// converts source to ierror structure, if possible
-	s, ok := target.(*ierror)
-	if !ok {
-		return false
-	}
-
-	return s.code&t.code > 0
-}
+import (
+	"errors"
+)
 
 // TODO add better description
 // HasCode Compares error with error code
@@ -34,4 +12,13 @@ func HasCode(err error, code ErrCode) bool {
 		return false
 	}
 	return (code & e.code) > 0
+}
+
+// TODO add better description
+// Is Compares errors
+func Is(source, target error) bool {
+	// ierrors uses the func (ie *ierror) Is(err error) bool
+	// to avaliate the two errors if they are an ierror type.
+	// If not they will execute the standard comparison.
+	return errors.Is(source, target)
 }
